Several adjustments made to add more info for config and move some functions around
